Refactor CheckScanner to CheckModule
