+++ lib/msf/core/exploit/mixins.rb
@@ -4,10 +4,10 @@
 #
 
 # Behavior
+require 'msf/core/exploit/check_module'
 require 'msf/core/exploit/brute'
 require 'msf/core/exploit/brutetargets'
 require 'msf/core/exploit/browser_autopwn'
-require 'msf/core/exploit/check_scanner'
 
 # Payload
 require 'msf/core/exploit/egghunter'
